implemented and used my_rand in order to have the same sequence on all platforms
still not the same results on all platform

+++ include/constants.h
@@ -15,7 +15,7 @@ public:
 
 	double *constants_double;
 
-	long constants_type;
+	int constants_type;
 	bool constants_can_evolve;
 	double constants_mutation_max_deviation;
